RFCT: extract componnets into summary bar and import it
