+++ frontend/src/features/summary-bar/subfeatures/utility-panel/UtilityPanel.jsx
@@ -1,0 +1,16 @@
+import NavSheet from "@/components/widgets/nav-sheet/NavSheet";
+import { navigation_info } from "@/constants/navigation";
+
+const UtilityPanel = () => {
+  return (
+    <NavSheet
+      nav_info={navigation_info.filter(
+        (nav) => nav.name.toLowerCase() != "cmt",
+      )}
+    >
+      <div className="hw-full flex-center"></div>
+    </NavSheet>
+  );
+};
+
+export default UtilityPanel;
